abstract out some tests, shuffle around maria-specific stuff
